% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_utils.R
\name{add_alpha}
\alias{add_alpha}
\title{Makes Color(s) Transparent}
\usage{
add_alpha(color, alpha = 1)
}
\arguments{
\item{color}{Color vector: vector of any of the three kinds of R color
specifications, i.e., either a color name (as listed by
\code{\link{colors}()}), a hexadecimal string of the form \code{"#rrggbb"}
or \code{"#rrggbbaa"} (see \code{\link{rgb}}), or a positive integer
\code{i} meaning \code{\link{palette}()[i]}.}

\item{alpha}{Numeric: amount of transparency in `[0, 1]`}
}
\value{
A character vector of HEX colors of same length as \code{color}
}
\examples{
add_alpha(c('red', 'blue'), alpha = 0.5)
}
\seealso{
\code{\link[ggplot2]{scale_alpha}}
}
