% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stat_utils.R
\name{anova.test}
\alias{anova.test}
\title{Fit an Analysis of Variance Model}
\usage{
\method{anova}{test}(...)
}
\description{
Same as code{\link[stats]{aov}} with only one difference that
 p value of the test is added to the result as a component named 'p.value'.
 Please refer to \code{\link[stats]{aov}} for more details
}
