% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_utils.R
\name{base_log_breaks}
\alias{base_log_breaks}
\title{Compute Pretty Axis Ticks of Log-scale}
\usage{
base_log_breaks(n = 10)
}
\arguments{
\item{n}{Integer: value indicating (approximately) the desired number of
intervals. See also \code{nint} in \code{\link[grDevices]{axisTicks}}}
}
\value{
A function that taks a numeric vector and computes log-scale axis
 tick locations
}
\description{
A wrapper function of \code{\link[grDevices]{axisTicks}} that
 uses log scale.
}
\examples{
tick_func <- base_log_breaks(n = 10)
tick_func(runif(100, 0, 10))
}
\seealso{
\code{\link[grDevices]{axisTicks}}
}
