% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_utils.R
\name{gg_color_hue}
\alias{gg_color_hue}
\title{Emulates ggplot2 Default Color Palette}
\usage{
gg_color_hue(n)
}
\arguments{
\item{n}{Integer: number of output colors}
}
\value{
A character vector of HEX colors of length \code{n}
}
\examples{
gg_color_hue(n = 3)
}
