% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gg_scatter.R
\name{gg_scatter}
\alias{gg_scatter}
\title{Scatter plots}
\usage{
gg_scatter(data, x, y, label = NULL, facet_r = NULL, facet_c = NULL,
  facet_r_levels = NULL, facet_c_levels = NULL, facet_scale = "free",
  facet_space = "free", color_var = NULL, all_colors = NULL,
  color_lab = color_var, shape_var = NULL, all_shapes = NULL,
  shape_lab = shape_var, add_label = FALSE, is_repel = FALSE,
  label_xlim = c(-Inf, Inf), label_ylim = c(-Inf, Inf),
  label_xloc = "middle", label_yloc = "middle", x_lab = x, y_lab = y,
  title = "", x_limit = NULL, y_limit = NULL, x_log = FALSE,
  y_log = FALSE, add_legend = TRUE, legend_pos = "bottom",
  reference_hline = NULL, reference_vline = NULL, bw_theme = TRUE,
  grids = "on")
}
\arguments{
\item{data}{Data frame: default dataset to use for plot}

\item{x}{Character: name of a \code{data} column mapped to x-axis
variable, i.e. time}

\item{y}{Character: name of a \code{data} column mapped to y-axis
variable}

\item{label}{Character: name of a \code{data} column mapped to point labels}

\item{facet_r}{Character: name of a \code{data} column mapped to the facet
row in panel plot layout. Check \code{\link[ggplot2]{facet_grid}} for more
details}

\item{facet_c}{Character: name of a \code{data} column mapped to the facet
column in panel plot layout. Check \code{\link[ggplot2]{facet_grid}} for
more details}

\item{facet_r_levels}{Vector/List: a named vector/list that specifies the
levels and labels of \code{facet_r}}

\item{facet_c_levels}{Vector/List: a named vector/list that specifies the
levels and labels of \code{facet_c}}

\item{facet_scale}{Character: Are scales shared across all facets. Refer to
the `scale` argument in \code{\link[ggplot2]{facet_grid}}. Default `free`
means that scales are not shared}

\item{facet_space}{Character: Refer to the `space` argument in
\code{\link[ggplot2]{facet_grid}}. Default `free` means both height and
width will vary}

\item{color_var}{Character: name of a \code{data} column mapped to the color
of the points}

\item{all_colors}{Vector: a vector of valid color representations for all
the point colors. See \code{\link{is.color}} for valid color definition}

\item{color_lab}{Character: title of color legend}

\item{shape_var}{Character name of a \code{data} column mapped to the shape
of the points}

\item{all_shapes}{Vector: a vector of all the point shapes. See
\href{http://sape.inf.usi.ch/quick-reference/ggplot2/shape}{ggplot2 Quick Reference: shape}}

\item{shape_lab}{Character: title of shape legend}

\item{add_label}{Logical: whether or not to add labels for the points
Default is set to `FALSE`}

\item{is_repel}{Logical: whether or not to avoid text overlapping using
\href{https://cran.r-project.org/web/packages/ggrepel/vignettes/ggrepel.html}{ggrepel}}

\item{label_xlim}{Numeric vector of length 2: it defines x-axis range to
select points for labelling. See \code{label_xloc} for more details of point
labelling rule. Default `c(-Inf, Inf)` includes all the points over x-axis}

\item{label_ylim}{Numeric vector of length 2: it defines y-axis range to
select points for labelling. See \code{label_yloc} for more details of point
labelling rule. Default `c(-Inf, Inf)` includes all the points over x-axis}

\item{label_xloc}{Character: Either 'middle' or 'sides'. Together with
\code{label_xlim}, it defines the rule of point labelling. Only points,
whose x-axis coordinates fall within (if 'middle') or beyond (if 'sides')
the x-axis range defined by \code{label_xlim}, are to be labelled.}

\item{label_yloc}{Character: Either 'middle' or 'sides'. Together with
\code{label_ylim}, it defines the rule of point labelling. Only points,
whose y-axis coordinates fall within (if 'middle') or beyond (if 'sides')
the y-axis range defined by \code{label_ylim}, are to be labelled.}

\item{x_lab}{Character: x-axis label}

\item{y_lab}{Character: y-axis label}

\item{title}{Character: barplot title}

\item{x_limit}{Numeric vector of length 2: limits for x-axis, e.g.
\code{c(0, 10)}.}

\item{y_limit}{Numeric vector of length 2: limits for y-axis, e.g.
\code{c(-5, 5)}}

\item{x_log}{Logical: whether or not to use log scale for x-axis. Default is
set to `TRUE`}

\item{y_log}{Logical: whether or not to use log scale for y-axis. Default is
set to `TRUE`}

\item{add_legend}{Logical: \code{TRUE} (default) to show legend and
\code{FALSE} otherwise}

\item{legend_pos}{Character: dictates the location where to place the legend.
By default, the legend will be place beneath the actual plot}

\item{reference_hline}{Numeric vector: locations of horizontal reference
line(s) if there is any}

\item{reference_vline}{Numeric vector: locations of vertical reference
line(s) if there is any}

\item{bw_theme}{Logical: If `TRUE` (default), black-and-white theme will be
used. Refer to \code{\link[ggplot2]{theme_bw}} for more details}

\item{grids}{Character: grids option. Must be one of `c('on', 'major',
'off')` with 'on' having both major and minor grids, 'major' having only
major grids, and 'off' having no grids}
}
\value{
An object of class ggplot. Can be directly sent to plot with
 \code{\link{print}}
}
\description{
Creates a scatter plot
}
\details{
This function relies on \href{http://ggplot2.org/}{ggplot2} package to
create a scatter plot with ability to label points within specific region.
Point labelling has the option to repel text overlapping with
\href{https://cran.r-project.org/web/packages/ggrepel/vignettes/ggrepel.html}{ggrepel}.
The function supports panel plot layout. By default black-and-white theme is
used.
}
\examples{
data <- mtcars
data$model <- row.names(data)
data$cyl <- factor(data$cyl)
gg_scatter(data, x = 'wt', y = 'mpg', label = 'model', color_var = 'cyl',
           color_lab = 'Cylinder', add_label = TRUE, is_repel = TRUE,
           label_xlim = c(2, 5), label_xloc = 'sides',
           x_lab = 'Weight (1000 lbs)', y_lab = 'Miles/(US) gallon')

gg_scatter(data, x = 'wt', y = 'mpg', label = 'model', facet_c = 'cyl',
           facet_c_levels = c('4 Cylinders' = '4', '6 Cylinders' = '6',
                              '8 Cylinders' = '8'),
           add_label = TRUE, is_repel = TRUE,
           x_lab = 'Weight (1000 lbs)', y_lab = 'Miles/(US) gallon')

}
\author{
Feiyang Niu (Feiyang.Niu@gilead.com)
}
