% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gg_time_profiling.R
\name{gg_time_profiling}
\alias{gg_time_profiling}
\title{Time-profiling Plots}
\usage{
gg_time_profiling(data, x, y, subject = NULL, group = NULL,
  group_levels = NULL, facet_r = NULL, facet_c = NULL,
  facet_r_levels = NULL, facet_c_levels = NULL, facet_scale = "free",
  facet_space = "free", x_lab = x, y_lab = y, group_lab = group,
  title = "", x_limit = NULL, y_limit = NULL, x_tick_label = x,
  all_xticks = FALSE, x_tick_angle = 0, y_tick_angle = 0,
  geoms = c("point", "line", "sumline", "boxplot"), avg_method = "mean",
  var_method = "se", y_log = FALSE, sample_size = TRUE,
  reference_hline = NULL, reference_vline = NULL, subject_show = FALSE,
  subject_label_size = 0.8, add_legend = TRUE, legend_pos = "bottom",
  all_colors = NULL, all_linetypes = NULL, randseed = 0,
  return_data = FALSE)
}
\arguments{
\item{data}{Data frame: default dataset to use for plot}

\item{x}{Character: name of a \code{data} column mapped to x-axis
variable, i.e. time}

\item{y}{Character: name of a \code{data} column mapped to y-axis
variable}

\item{subject}{Character: name of a \code{data} column mapped subject IDs
or other types of IDs that define membership of repeated measurements in
the data}

\item{group}{Character: name of a \code{data} column mapped to the fill of
bars}

\item{group_levels}{Vector: levels of \code{group} column}

\item{facet_r}{Character: name of a \code{data} column mapped to the facet
row in panel plot layout. Check \code{\link[ggplot2]{facet_grid}} for more
details}

\item{facet_c}{Character: name of a \code{data} column mapped to the facet
column in panel plot layout. Check \code{\link[ggplot2]{facet_grid}} for
more details}

\item{facet_r_levels}{Vector/List: a named vector/list that specifies the
levels and labels of facet row.}

\item{facet_c_levels}{Vector/List: a named vector/list that specifies the
levels and labels of facet column.}

\item{facet_scale}{Character: Are scales shared across all facets. Refer to
the `scale` argument in \code{\link[ggplot2]{facet_grid}}. Default `free`
means that scales are not shared}

\item{facet_space}{Character: Refer to the `space` argument in
\code{\link[ggplot2]{facet_grid}}. Default `free` means both height and
width will vary}

\item{x_lab}{Character: x-axis label}

\item{y_lab}{Character: y-axis label}

\item{group_lab}{Character: group variable label}

\item{title}{Character: barplot title}

\item{x_limit}{Numeric vector of length 2: limits for x-axis, e.g.
\code{c(0, 10)}.}

\item{y_limit}{Numeric vector of length 2: limits for y-axis, e.g.
\code{c(-5, 5)}}

\item{x_tick_label}{Vector: x-axis tick marks}

\item{all_xticks}{Logical: \code{TRUE} (default) to display all marks
specified as the unique values of \code{x_tick_label}}

\item{x_tick_angle}{Numeric: the orientation angle (in [0, 360]) of the
x-axis tick marks. By default, the label will be horizontal.}

\item{y_tick_angle}{Numeric: the orientation angle (in [0, 360]) of the
y-axis tick marks. By default, the label will be horizontal.}

\item{geoms}{Character vector: specify geometric components to be included
in the plot. For example, if the user needs to create a boxplot with
raw data points, specify \code{geoms = c('boxplot', 'point')}; if the user
wants to create a Mean + SE plot, let \code{geoms = c('sumline', 'point')}
together with specification of \code{avg_method} and \code{var_method}}

\item{avg_method}{Character: dicates whether to use \code{\link[base]{mean}}
(\code{avg_method = 'mean'}) or \code{\link[stats]{median}}. Be noted that
only these two options are permitted and both methods will remove \code{NA}s
in calculation}

\item{var_method}{Character: dicates whether to use standard error
(\code{var_method = 'se'}) or standard deviation (\code{var_method = 'sd'}).
Be noted that only these two options are permitted and both methods will
remove \code{NA}s in calculation}

\item{y_log}{Logical: \code{TRUE} to use log-scale for y-axis and
\code{FALSE} (default) otherwise.}

\item{sample_size}{Logical: \code{TRUE} (default) to place sample size
annotation along x-axis tick marks and \code{FALSE} otherwise}

\item{reference_hline}{Numeric vector: locations of horizontal reference
line(s) if there is any}

\item{reference_vline}{Numeric vector: locations of vertical reference
line(s) if there is any}

\item{subject_show}{Logical: \code{TRUE} to show subject IDs (or other types)
and \code{FALSE} (default) otherwise}

\item{add_legend}{Logical: \code{TRUE} (default) to show legend and
\code{FALSE} otherwise}

\item{legend_pos}{Character: dictates the location where to place the legend.
By default, the legend will be place beneath the actual plot}

\item{all_colors}{Vector of valid color representation: when \code{group} is
present, categories of \code{group} will take different color for specified
geometric objects. This argument allows the user to overwrite default color
theme. See \code{col} argument in \code{\link[grDevices]{col2rgb}} for types
of valid color representation}

\item{all_linetypes}{Numeric vector: by default, lines of different
categories of \code{group}, when present, will be solid lines. This argument
allows the user to overwrite the default line types. See
\href{http://sape.inf.usi.ch/quick-reference/ggplot2/linetype}{linetypes}
for valid ggplot2 linetype specification}

\item{randseed}{Numeric: random seed can be set in producing jittered points
when \code{geoms} includes `point`}

\item{return_data}{Logical: \code{TRUE} to return both plot and the data
used to produce the plot and \code{FALSE} (default) to return only plot}
}
\value{
An object of class ggplot (if \code{return_data = FALSE}) or a list
 of two components: an object of class ggplot and a data frame that used to
 generate the plot
}
\description{
Creats time-profiling plots with
\href{https://en.wikipedia.org/wiki/Panel_data}{longitudinal data}
}
\details{
This function relies on \href{http://ggplot2.org/}{ggplot2} package to
create time-profiling plots including: Mean + SD/SE plot, Mean + Median
plot, Boxplot, Spaghetti plot and their mixture. Panel plot layout is
supported. By default black-and-white theme is used.
}
\examples{
library(CommonPlots)
data(koopLME)

# Mean + SE plot
data <- koopLME
data$siblings <- factor(with(data, ifelse(SIBLINGS >= 3, '>= 3', SIBLINGS)),
                        levels = c(0, 1, 2, '>= 3'))
mean_se_plot <- gg_time_profiling(
    data, x = 'TIMETRND', y = 'LOGWAGE', group = 'siblings',
    geoms = c('sumline'), avg_method = 'mean',
    var_method = 'se', x_lab = 'Time trend',
    y_lab = 'Log of hourly wage', group_lab = 'Number of siblings',
    title = 'Mean + SE plot'
)

# Median + IQR plot
data <- koopLME
data$siblings <- factor(with(data, ifelse(SIBLINGS >= 3, '>= 3', SIBLINGS)),
                        levels = c(0, 1, 2, '>= 3'))
sib_labels <- paste('Number of siblings:', levels(data$siblings))
median_iqr_plot <- gg_time_profiling(
    data, x = 'TIMETRND', y = 'LOGWAGE', facet_c = 'siblings',
    facet_c_levels = setNames(levels(data$siblings), sib_labels),
    geoms = c('sumline'), avg_method = 'median',
    x_lab = 'Time trend', y_lab = 'Log of hourly wage',
    title = 'Median + IQR plot'
)

# Spaghetti plot
data <- koopLME
data <- with(data, data[PERSONID \%in\% unique(PERSONID)[1:100], ])
spaghetti_plot <- gg_time_profiling(
    data, x = 'TIMETRND', y = 'LOGWAGE', subject = 'PERSONID',
    geoms = c('line'), x_lab = 'Time trend',
    y_lab = 'Log of hourly wage', title = 'Spaghetti plot',
    subject_show = TRUE
)

# Box plot over time
boxplot <- gg_time_profiling(
    koopLME, x = 'TIMETRND', y = 'LOGWAGE',
    geoms = c('boxplot', 'point'), x_lab = 'Time trend',
    y_lab = 'Log of hourly wage', title = 'Box plot',
    sample_size = FALSE, all_xticks = TRUE
)


}
\author{
Feiyang Niu (Feiyang.Niu@gilead.com)
}
