% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_utils.R
\name{gg_wrapper}
\alias{gg_wrapper}
\title{Set/Modify ggplot2 Components}
\usage{
gg_wrapper(..., facet_r = NULL, facet_c = NULL, facet_scale = "fixed",
  facet_space = "fixed", is_x_continuous = TRUE, is_y_continuous = TRUE,
  x_lab = NULL, y_lab = NULL, title = NULL, x_limit = NULL,
  y_limit = NULL, x_log = FALSE, y_log = FALSE, x_reverse = FALSE,
  y_reverse = FALSE, x_axis_breaks = NULL, y_axis_breaks = NULL,
  x_tick_labels = x_axis_breaks, y_tick_labels = y_axis_breaks,
  x_tick_angle = 0, y_tick_angle = 0, x_hjust = 0.5, y_hjust = 0.5,
  add_legend = TRUE, legend_pos = "bottom", color_var = NULL,
  color_lab = color_var, all_colors = NULL, fill_var = NULL,
  fill_lab = fill_var, all_fills = NULL, size_var = NULL,
  size_lab = size_var, all_sizes = NULL, shape_var = NULL,
  shape_lab = shape_var, all_shapes = NULL, linetype_var = NULL,
  linetype_lab = linetype_var, all_linetypes = NULL, alpha_var = NULL,
  alpha_lab = alpha_var, all_alphas = NULL, reference_hline = NULL,
  reference_vline = NULL, grids = "on", bw_theme = TRUE,
  strip_background_color = NULL, x_expand = NULL, y_expand = NULL)
}
\description{
Helper function that allows the user to easily set/modify
 ggplot2 object components, such as title, x-/y-axis log scale, grids and
 etc..
}
