% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_utils.R
\name{is.color}
\alias{is.color}
\title{Check If Valid Color Representation}
\usage{
is.color(obj)
}
\arguments{
\item{obj}{Object: object to be tested}
}
\value{
Logical: \code{TRUE} if the object is a valid color representation,
 \code{FALSE} otherwise
}
\description{
A valid color representation is any of the three kinds of R
 color specifications, i.e., either a color name (as listed by
 \code{\link{colors}()}), a hexadecimal string of the form \code{"#rrggbb"}
 or \code{"#rrggbbaa"} (see \code{\link{rgb}}), or a positive integer
 \code{i} meaning \code{\link{palette}()[i]}.
}
\examples{
is.color(c('red', 'blue'))
is.color(1)
is.color('abcd')
}
\seealso{
\code{\link[grDevices]{col2rgb}}
}
