% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stat_utils.R
\name{is.formula}
\alias{is.formula}
\title{Test if an object is a formula}
\usage{
is.formula(x)
}
\arguments{
\item{x}{an object}
}
\value{
\code{TRUE} if \code{x} is a formula and \code{FALSE} otherwise
}
\examples{
is.formula(weight ~ height)

}
\author{
Feiyang Niu (Feiyang.Niu@gilead.com)
}
