% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stat_utils.R
\name{JonckheereTerpstraTest}
\alias{JonckheereTerpstraTest}
\title{Exact Version of Jonckheere-Terpstra Test}
\usage{
JonckheereTerpstraTest(...)
}
\description{
Jonckheere-Terpstra test to test for ordered differences among
 classes. Please refer to \code{\link[DescTools]{JonckheereTerpstraTest}} for
 more details
}
