% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stat_utils.R
\name{mean_na}
\alias{mean_na}
\title{Arithmetic Mean with NAs Removed}
\usage{
mean_na(x, ...)
}
\arguments{
\item{x}{See \code{x} in \code{\link[base]{mean}}}

\item{...}{Ellipsis: arguments passed to \code{\link[base]{mean}}}
}
\value{
See return of \code{\link[base]{mean}}
}
\description{
\code{\link[base]{mean}} function with \code{na.rm = TRUE}
}
\examples{
mean_na(c(NA, 9, 10))
}
\seealso{
\code{\link[base]{mean}}
}
