% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stat_utils.R
\name{mean_sd}
\alias{mean_sd}
\title{Calculate Mean And Standard Deviation}
\usage{
mean_sd(x)
}
\arguments{
\item{x}{Numeric vector}
}
\value{
A data frame with columns y, ymin, and ymax
}
\description{
For use with \code{\link[ggplot2]{stat_summary}}
}
\examples{
mean_sd(c(NA, 1:100))
}
\seealso{
\code{\link[ggplot2]{mean_se}}
}
