% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stat_utils.R
\name{median_iqr}
\alias{median_iqr}
\title{Calculate Median And IQR}
\usage{
median_iqr(x, type = 2)
}
\arguments{
\item{x}{Numeric vector}

\item{type}{Integer: an integer between 1 and 9. See \code{type} in
\code{\link[stats]{quantile}}}
}
\value{
A data frame with columns y, ymin, and ymax
}
\description{
For use with \code{\link[ggplot2]{stat_summary}}
}
\examples{
median_iqr(c(NA, 1:100))
}
\seealso{
\code{\link[ggplot2]{mean_se}}
}
