% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_utils.R
\name{reverselog_trans}
\alias{reverselog_trans}
\title{Create A Log-scale And Reverse Order Transformation Object}
\usage{
reverselog_trans(base = 10)
}
\arguments{
\item{base}{Integer: the base with respect to which logarithms are computed}
}
\value{
A transformation object
}
\description{
A transformation created with \code{\link[scales]{trans_new}}
 by setting \code{transform} to log-scale and \code{inverse} to
 reverse-order function
}
\examples{
\dontrun{
revlog_trans <- reverselog_trans(base = 10)
}
}
\seealso{
\code{\link[scales]{trans_new}}
}
