% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stat_utils.R
\name{ternary}
\alias{ternary}
\title{R version ternary function}
\usage{
ternary(expr, obj_x, obj_y)
}
\arguments{
\item{obj_x}{Object x}

\item{obj_y}{Object y}
}
\value{
\code{obj_x} if \code{expr} is evaluated \code{TRUE} and
 \code{obj_y} otherwise
}
\examples{
ternary(TRUE, list(), list(a = 10))
}
