% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gg_barplot.R
\name{gg_barplot}
\alias{gg_barplot}
\title{Bar plots}
\usage{
gg_barplot(data, var, var_levels = NULL, group = NULL, facet_r = NULL,
  facet_c = NULL, facet_scale = "free", facet_space = "free",
  x_lab = NULL, y_lab = NULL, title = NULL, add_counts = TRUE,
  counts_pos = "inside-bar", bar_label_align = "left",
  bar_label_angle = NULL, group_bar_position = "dodge", grids = "on",
  bw_theme = TRUE, horizontal = TRUE)
}
\arguments{
\item{data}{Data frame: default dataset to use for plot}

\item{var}{Character: name of a \code{data} column whose counts provided as
the height of bars}

\item{var_levels}{Character vector: levels of \code{var}}

\item{group}{Character: name of a \code{data} column mapped to the fill of
bars}

\item{facet_r}{Character: name of a \code{data} column mapped to the facet
row in panel plot layout. Check \code{\link[ggplot2]{facet_grid}} for more
details}

\item{facet_c}{Character: name of a \code{data} column mapped to the facet
column in panel plot layout. Check \code{\link[ggplot2]{facet_grid}} for
more details}

\item{facet_scale}{Character: Are scales shared across all facets. Refer to
the `scale` argument in \code{\link[ggplot2]{facet_grid}}. Default `free`
means that scales are not shared}

\item{facet_space}{Character: Refer to the `space` argument in
\code{\link[ggplot2]{facet_grid}}. Default `free` means both height and
width will vary}

\item{x_lab}{Character: x-axis label}

\item{y_lab}{Character: y-axis label}

\item{title}{Character: barplot title}

\item{add_counts}{Logical: If `TRUE` (default), counts will be added to the
bars}

\item{counts_pos}{Character: position to add counts. If `inside-bar`
(default), counts will be inside the bars; if `outside-bar`, counts be
placed outside the bars. When \code{group} is present and
\code{group_bar_position} is specified as `stack`, counts will be placed
in the center of the bars}

\item{bar_label_align}{Character: alignment of the bar labels. If `left`
(default), labels will be left-aligned; if `center`, labels will be
center-aligned; if `right`, labels will be right-aligned}

\item{bar_label_angle}{Numeric: the orientation angle (in [0, 360]) of the
bar labels on the axis. By default, the label will be perpendicular to the
axis}

\item{group_bar_position}{Character: position adjustment for bars of same
group when \code{group} is present. If `dodge` (default), bars of same
group are placed side by side; if `stack`, bars of same group are stacked
one on top of another.}

\item{grids}{Character: control the grids. If `on` (default), grids will be
drawn; if `x`, only grids on x-axis will be drawn; if `y`, only grids on
y-axis will be drawn; if `off`, no grids will be drawn}

\item{bw_theme}{Logical: If `TRUE` (default), black-and-white theme will be
used. Refer to \code{\link[ggplot2]{theme_bw}} for more details}

\item{horizontal}{Logical: If `TRUE` (default), horizontal barplot is drawn;
if `FALSE`, vertical barplot is drawn}
}
\value{
An object of class ggplot. Can be directly sent to plot with
 \code{\link{print}}
}
\description{
Creates a bar plot with vertical or horizontal bars
}
\details{
This function relies on \href{http://ggplot2.org/}{ggplot2} package to
create a bar plot with vertical or horizontal bars. The function supports
panel plot layout. Counts information can be added to the corresponding
bars. By default black-and-white theme is used.
}
\examples{
gg_barplot(mpg, var = 'class', horizontal = FALSE)
gg_barplot(mpg, var = 'class', group = 'cyl')
gg_barplot(mpg, var = 'class', group = 'cyl', add_counts = FALSE)
gg_barplot(mpg, var = 'class', group = 'cyl', group_bar_position = 'stack')

}
\author{
Feiyang Niu (Feiyang.Niu@gilead.com)
}
