% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stat_utils.R
\name{JonckheereTerpstraTest}
\alias{JonckheereTerpstraTest}
\title{Exact Version of Jonckheere-Terpstra Test}
\usage{
JonckheereTerpstraTest(formula, data, ...)
}
\arguments{
\item{formula}{a formula of the form `y ~ group` where y gives the data
values and group the corresponding groups.}

\item{data}{a data.frame containing the variables in the formula `formula`}
}
\description{
Jonckheere-Terpstra test to test for ordered differences among
 classes.
}
